requirements.txt
discord.py
pillow
motor
aiohttp